requirements.txt
PyOpenGL==3.1.7
pygame==2.4.0
vedo==2023.4.4
Pillow==9.5.0
PyOpenGL-accelerate==3.1.7
